requirements.txt
certifi==2020.6.20
ConfigArgParse==1.2.3
cycler==0.10.0
future==0.18.2
h5py==2.10.0
kiwisolver==1.2.0
matplotlib==3.3.1
numpy==1.16.2
opencv-python==4.4.0.42
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.48.2
